Create Atm
This code implements an ATM system in Java. It consists of five classes - User, Account, Transaction, ATM, and a main class. When the program starts, the user is prompted to enter their ID and PIN. If the login is successful, the ATM functionalities are unlocked. The user can then perform transactions such as withdrawing money, depositing money, transferring money, checking their balance, or quitting the program. The program uses console input and output for user interaction.